lang_chain_explorations/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.7.22
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
dataclasses-json==0.6.2
distlib==0.3.7
distro==1.8.0
dulwich==0.21.6
fastapi==0.104.1
fastjsonschema==2.18.1
filelock==3.13.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
httpx-sse==0.3.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jsonpatch==1.33
jsonpointer==2.4
keyring==24.3.0
langchain==0.0.335
langchain-cli==0.0.16
langserve==0.0.27
langsmith==0.0.64
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.2.4
orjson==3.9.10
packaging==23.2
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.16.1
pyproject_hooks==1.0.0
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.6.0
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sse-starlette==1.6.5
starlette==0.27.0
tenacity==8.2.3
tomlkit==0.12.2
tqdm==4.66.1
trove-classifiers==2023.11.13
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.23.2
virtualenv==20.24.6
yarl==1.9.2
zipp==3.17.0
